Undo module rename
